// apps/client/src/app/core/swagger/model/createComment.ts
/**
 * NgReports
 * NG Reports API
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface CreateComment { 
    /**
     * The ID of the report to which this comment belongs.
     */
    reportId: string;
    /**
     * The content of the comment.
     */
    content: string;
}