// apps/client/src/app/core/swagger/model/updateComment.ts
/**
 * NgReports
 * NG Reports API
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface UpdateComment {
  /**
   * The ID of the report to which this comment belongs.
   */
  reportId?: string;
  /**
   * The new content of the comment.
   */
  content?: string;
}